requirements.txt
click>=8.0.0
psycopg2-binary>=2.9.0
requests>=2.28.0
python-dotenv>=0.19.0
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0
alembic>=1.10
